Move the bulk of the direwolf main back into the direwolf package
Issues using cgo structs across packages - C.struct_audio_s vs
direwolf.C.struct_audio_s - it seems massively easier if the cmds don't
import C themselves
